+++ fabric/network.py
@@ -449,7 +449,8 @@ def connect(user, host, port, cache, seek_gateway=True):
                 timeout=env.timeout,
                 allow_agent=not env.no_agent,
                 look_for_keys=not env.no_keys,
-                sock=sock
+                sock=sock,
+                banner_timeout=env.banner_timeout
             )
             connected = True